requirements.txt
numpy==1.22.0
pandas==1.5.3
Pillow==10.1.0
Pillow==10.3.0
scikit_learn==1.3.0
torch==2.1.0
torchvision==0.16.0
